refactor: database models + docker compose
