+++ backend/app/models/tasks/task_types.py
@@ -1,0 +1,6 @@
+from app.models.base import Base
+from sqlalchemy import Column, String, Text
+
+class TaskType(Base):
+    name = Column(String(255), nullable=False)
+    description = Column(Text)
