+++ backend/app/models/tasks/volunteer_tasks.py
@@ -1,0 +1,10 @@
+from sqlalchemy import Column, String, Text, Integer, ForeignKey
+from app.models.base import Base
+
+class VolunteerTask(Base):
+    issue_id = Column(Integer, ForeignKey("issue.id"))
+    title = Column(String(255), nullable=False)
+    description = Column(Text)
+    status = Column(String(50), nullable=False, default="PENDING")
+    urgency = Column(String(50))
+    complexity = Column(String(50))
