+++ docker-compose.yml
@@ -1,0 +1,60 @@
+version: '3.8'
+
+services:
+  db:
+    container_name: db
+    image: postgres:16-alpine
+    ports:
+      - '5433:5432'
+    environment:
+      POSTGRES_USER: ${DB_USER}
+      POSTGRES_PASSWORD: ${DB_PASSWORD}
+      POSTGRES_DB: ${DB_NAME}
+    volumes:
+      - db_data:/var/lib/postgresql/data
+
+  pgadmin:
+    container_name: pgadmin
+    image: dpage/pgadmin4
+    environment:
+      PGADMIN_DEFAULT_EMAIL: ${PG_EMAIL}
+      PGADMIN_DEFAULT_PASSWORD: ${PG_PASSWORD}
+    ports:
+      - '8081:80'
+    depends_on:
+      - db
+
+  backend:
+    container_name: backend
+    build:
+      context: ./backend
+    environment:
+      DATABASE_URL: postgresql://${DB_USER}:${DB_PASSWORD}@db:5432/${DB_NAME}
+      SECRET_KEY: ${SECRET_KEY}
+      ALGORITHM: ${ALGORITHM}
+      ACCESS_TOKEN_EXPIRE_MINUTES: ${ACCESS_TOKEN_EXPIRE_MINUTES}
+      ENVIRONMENT: ${ENVIRONMENT}
+      DEBUG: ${DEBUG}
+      UPLOAD_FOLDER: ${UPLOAD_FOLDER}
+      AI_MODEL_PATH: ${AI_MODEL_PATH}
+      AI_CONFIG_PATH: ${AI_CONFIG_PATH}
+      LOG_LEVEL: ${LOG_LEVEL}
+      CLIENT_URL: ${CLIENT_URL}
+      SERVER_PORT: ${SERVER_PORT}
+    depends_on:
+      - db
+    expose:
+      - '8080'
+
+  nginx:
+    container_name: nginx
+    build:
+      context: .
+      dockerfile: ./nginx/Dockerfile
+    depends_on:
+      - backend
+    ports:
+      - '80:80'
+
+volumes:
+  db_data:
